+++ R/jsSurveyGadget.R
@@ -289,7 +289,9 @@ jsSurveyGadget <- function(data, nfactor.limit = 20) {
         out <- out[!is.na(get(input$weights_vname))]
       }
 
-      surveydata <- survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out)
+      #surveydata <- survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out)
+      surveydata <- tryCatch(survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out),
+                             error = function(e){return(survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out, nest = T))})
 
       out.label <- mk.lev(out)
 
@@ -302,7 +304,9 @@ jsSurveyGadget <- function(data, nfactor.limit = 20) {
 
           if (input$var_subset %in% var.factor){
             out <- out[get(input$var_subset) %in% input$val_subset]
-            surveydata <- survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out)
+            #surveydata <- survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out)
+            surveydata <- tryCatch(survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out),
+                                   error = function(e){return(survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out, nest = T))})
             #var.factor <- c(data()$factor_original, input$factor_vname)
             out[, (var.factor) := lapply(.SD, factor), .SDcols = var.factor]
             out.label2 <- mk.lev(out)[, c("variable", "class", "level")]