+++ NEWS.md
@@ -1,3 +1,7 @@
+# jsmodule 0.9.2
+
+* Bug fixes: Apply nested survey design to `jsSurveyGadget` and `FileSurveyInput`
+
 # jsmodule 0.9.1
 
 ## Bug fixes